+++ mixi-SwiftUI-Training/mixi-SwiftUI-Training/Views/RepoDetailView.swift
@@ -1,0 +1,20 @@
+//
+//  RepoDetailView.swift
+//  mixi-SwiftUI-Training
+//
+//  Created by 梅原 奈輝 on 2022/06/11.
+//
+
+import SwiftUI
+
+struct RepoDetailView: View {
+    var body: some View {
+        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
+    }
+}
+
+struct RepoDetailView_Previews: PreviewProvider {
+    static var previews: some View {
+        RepoDetailView()
+    }
+}
